+++ tests/effects/teffects2.nim
@@ -4,7 +4,7 @@ discard """
 """
 {.push warningAsError[Effect]: on.}
 type
-  TObj = object {.pure, inheritable.}
+  TObj {.pure, inheritable.} = object
   TObjB = object of TObj
     a, b, c: string