+++ tests/types/tremoved_pragma_syntax.nim
@@ -1,0 +1,5 @@
+discard """
+  errormsg: "invalid indentation"
+"""
+
+type Foo = object {.final.} # deprecated since v0.20.0